Properly separate enumeratum and enumeration tests
